feat: it contains the first configuration of gg-react/gg-ui project
It provides a initial configuration for gg-ui React library in Typescript using Rollup to bundle the
source code into cjs and esm files, Jest for unit tests, Commitlint for commit patterns and Husky to
lock quality commits. It also have a markdown and two ci files of Github actions, one triggered on
pullrequests to main branch and other that builds the aplication and publish it into npm in every
release.

"feat #1"